refactor: выделить RootLayout и страницы
