+++ package.json
@@ -16,6 +16,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hot-toast": "^2.5.2",
+    "react-router-dom": "^6.22.3",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
