+++ src/pages/InventoryPage.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import RootLayout from '../components/RootLayout';
+
+export default function InventoryPage() {
+  return <RootLayout />;
+}
